requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
apache-airflow==2.10.5
apache-airflow-providers-celery==3.5.1
apache-airflow-providers-common-compat==1.2.0
apache-airflow-providers-common-io==1.2.0
apache-airflow-providers-common-sql==1.10.0
apache-airflow-providers-fab==1.4.0
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-http==4.8.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-smtp==1.8.0
apache-airflow-providers-sqlite==3.7.0
apispec==6.4.0
argcomplete==3.2.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
backports.zoneinfo==0.2.1
billiard==4.2.0
blinker==1.7.0
cachelib==0.9.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.0
croniter==6.0.0
cryptography==41.0.7
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.4.2
docutils==0.20.1
email-validator==1.3.1
eval-type-backport==0.2.2
exceptiongroup==1.2.0
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
flask-session==0.5.0
Flask-SQLAlchemy==2.5.1
flask-wtf==1.2.1
flower==2.0.1
frozenlist==1.4.1
fsspec==2023.12.2
google-re2==1.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.70.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.9.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==5.13.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kombu==5.3.5
lazy-object-proxy==1.10.0
limits==3.7.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.3.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.5.0
multidict==6.0.4
numpy==1.24.4
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
pathspec==0.12.1
pendulum==3.0.0
pkgutil-resolve-name==1.3.10
pluggy==1.5.0
prison==0.2.1
prometheus-client==0.19.0
prompt-toolkit==3.0.43
propcache==0.2.0
protobuf==5.29.4
psutil==5.9.7
psycopg2==2.9.10
pycparser==2.21
pydantic==2.5.3
pydantic-core==2.14.6
pygments==2.17.2
PyJWT==2.8.0
PySocks==1.7.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.7.0
rich-argparse==1.4.0
rpds-py==0.17.1
selenium==4.27.1
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.51
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.13.0
tornado==6.4
trio==0.27.0
trio-websocket==0.12.2
typing-extensions==4.9.0
tzdata==2023.4
uc-micro-py==1.0.2
unicodecsv==0.14.1
universal-pathlib==0.2.6
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.16.0
wsproto==1.2.0
wtforms==3.1.2
yarl==1.9.4
zipp==3.17.0
